Move AppContainer to pages folder
